+++ packages/liveblocks-core/src/client.ts
@@ -137,6 +137,7 @@ export type EnterOptions<P extends JsonObject = DP, S extends LsonObject = DS> =
  */
 export type PrivateClientApi<U extends BaseUserMeta, M extends BaseMetadata> = {
   readonly currentUserIdStore: Store<string | null>;
+  readonly mentionSuggestionsCache: Map<string, string[]>;
   readonly resolveMentionSuggestions: ClientOptions<U>["resolveMentionSuggestions"];
   readonly usersStore: BatchStore<U["info"] | undefined, string>;
   readonly roomsInfoStore: BatchStore<DRI | undefined, string>;
@@ -313,6 +314,58 @@ export type Client<U extends BaseUserMeta = DU, M extends BaseMetadata = DM> = {
    * Call this whenever you log out a user in your application.
    */
   logout(): void;
+
+  /**
+   * Advanced APIs related to the resolvers.
+   */
+  resolvers: {
+    /**
+     * Invalidate all or specific users that were previously cached by `resolveUsers`.
+     *
+     * @example
+     * // Invalidate all users
+     * client.resolvers.invalidateUsers();
+     *
+     * @example
+     * // Invalidate specific users
+     * client.resolvers.invalidateUsers(["user-1", "user-2"]);
+     *
+     * @example
+     * // Invalidate specific users using a predicate
+     * client.resolvers.invalidateUsers((user) => user.tags.includes("Engineering"));
+     */
+    invalidateUsers(
+      userIds?: string[] | ((user: NonNullable<U["info"]>) => boolean)
+    ): void;
+
+    /**
+     * Invalidate all or specific rooms info that were previously cached by `resolveRoomsInfo`.
+     *
+     * @example
+     * // Invalidate all rooms
+     * client.resolvers.invalidateRoomsInfo();
+     *
+     * @example
+     * // Invalidate specific rooms
+     * client.resolvers.invalidateRoomsInfo(["room-1", "room-2"]);
+     *
+     * @example
+     * // Invalidate specific rooms using a predicate
+     * client.resolvers.invalidateRoomsInfo((roomInfo) => roomInfo.url.includes("engineering"));
+     */
+    invalidateRoomsInfo(
+      roomIds?: string[] | ((info: NonNullable<DRI>) => boolean)
+    ): void;
+
+    /**
+     * Invalidate all mention suggestions cached by `resolveMentionSuggestions`.
+     *
+     * @example
+     * // Invalidate all mention suggestions
+     * client.resolvers.invalidateMentionSuggestions();
+     */
+    invalidateMentionSuggestions(): void;
+  };
 
   /**
    * @private
@@ -629,6 +682,12 @@ export function createClient<U extends BaseUserMeta = DU>(
   );
   const usersStore = createBatchStore(batchedResolveUsers);
 
+  function invalidateResolvedUsers(
+    userIds?: string[] | ((user: NonNullable<U["info"]>) => boolean)
+  ) {
+    usersStore.invalidate(userIds);
+  }
+
   const resolveRoomsInfo = clientOptions.resolveRoomsInfo;
   const warnIfNoResolveRoomsInfo = createDevelopmentWarning(
     () => !resolveRoomsInfo,
@@ -648,6 +707,18 @@ export function createClient<U extends BaseUserMeta = DU>(
   );
   const roomsInfoStore = createBatchStore(batchedResolveRoomsInfo);
 
+  function invalidateResolvedRoomsInfo(
+    roomIds?: string[] | ((info: NonNullable<DRI>) => boolean)
+  ) {
+    roomsInfoStore.invalidate(roomIds);
+  }
+
+  const mentionSuggestionsCache = new Map<string, string[]>();
+
+  function invalidateResolvedMentionSuggestions() {
+    mentionSuggestionsCache.clear();
+  }
+
   return Object.defineProperty(
     {
       enterRoom,
@@ -657,9 +728,17 @@ export function createClient<U extends BaseUserMeta = DU>(
 
       ...httpClientLike,
 
+      // Advanced resolvers APIs
+      resolvers: {
+        invalidateUsers: invalidateResolvedUsers,
+        invalidateRoomsInfo: invalidateResolvedRoomsInfo,
+        invalidateMentionSuggestions: invalidateResolvedMentionSuggestions,
+      },
+
       // Internal
       [kInternal]: {
         currentUserIdStore,
+        mentionSuggestionsCache,
         resolveMentionSuggestions: clientOptions.resolveMentionSuggestions,
         usersStore,
         roomsInfoStore,