+++ packages/liveblocks-core/src/lib/batch.ts
@@ -16,6 +16,14 @@ export type BatchCallback<O, I> = (
 export type BatchStore<O, I> = Observable<void> & {
   get: (input: I) => Promise<void>;
   getState: (input: I) => AsyncResult<O> | undefined;
+  invalidate: (inputs?: I[] | ((output: NonNullable<O>) => boolean)) => void;
+
+  /**
+   * @internal
+   *
+   * Only for testing.
+   */
+  _cacheKeys: () => string[];
 };
 
 interface Options {
@@ -170,6 +178,34 @@ export function createBatchStore<O, I>(batch: Batch<O, I>): BatchStore<O, I> {
     eventSource.notify();
   }
 
+  function invalidate(
+    inputs?: I[] | ((output: NonNullable<O>) => boolean)
+  ): void {
+    if (Array.isArray(inputs)) {
+      // Invalidate the specific calls.
+      for (const input of inputs) {
+        cache.delete(getCacheKey(input));
+      }
+    } else if (typeof inputs === "function") {
+      // Invalidate the predicated calls.
+      for (const [cacheKey, state] of cache.entries()) {
+        if (
+          state.data !== undefined &&
+          state.data !== null &&
+          inputs(state.data as NonNullable<O>)
+        ) {
+          cache.delete(cacheKey);
+        }
+      }
+    } else {
+      // Invalidate all calls.
+      cache.clear();
+    }
+
+    // Notify subscribers.
+    eventSource.notify();
+  }
+
   async function get(input: I): Promise<void> {
     const cacheKey = getCacheKey(input);
 
@@ -216,9 +252,20 @@ export function createBatchStore<O, I>(batch: Batch<O, I>): BatchStore<O, I> {
     return cache.get(cacheKey);
   }
 
+  /**
+   * @internal
+   *
+   * Only for testing.
+   */
+  function _cacheKeys() {
+    return [...cache.keys()];
+  }
+
   return {
     ...eventSource.observable,
     get,
     getState,
+    invalidate,
+    _cacheKeys,
   };
 }