+++ packages/liveblocks-react/src/liveblocks.tsx
@@ -666,24 +666,29 @@ function useUser_withClient<U extends BaseUserMeta>(
     [usersStore, userId]
   );
 
-  useEffect(() => {
-    // NOTE: .get() will trigger any actual fetches, whereas .getState() will not
-    void usersStore.get(userId);
-  }, [usersStore, userId]);
-
   const selector = useCallback(
     (state: ReturnType<typeof getUserState>) =>
       selectorFor_useUser(state, userId),
     [userId]
   );
 
-  return useSyncExternalStoreWithSelector(
+  const result = useSyncExternalStoreWithSelector(
     usersStore.subscribe,
     getUserState,
     getUserState,
     selector,
     shallow
   );
+
+  // Trigger a fetch if we don't have any data yet (whether initially or after an invalidation)
+  useEffect(() => {
+    if (!result.user && result.isLoading) {
+      // NOTE: .get() will trigger any actual fetches, whereas .getState() will not
+      void usersStore.get(userId);
+    }
+  }, [usersStore, userId, result]);
+
+  return result;
 }
 
 function useUserSuspense_withClient<U extends BaseUserMeta>(
@@ -743,17 +748,23 @@ function useRoomInfo_withClient(
     [roomId]
   );
 
-  useEffect(() => {
-    void roomsInfoStore.get(roomId);
-  }, [roomsInfoStore, roomId]);
-
-  return useSyncExternalStoreWithSelector(
+  const result = useSyncExternalStoreWithSelector(
     roomsInfoStore.subscribe,
     getRoomInfoState,
     getRoomInfoState,
     selector,
     shallow
   );
+
+  // Trigger a fetch if we don't have any data yet (whether initially or after an invalidation)
+  useEffect(() => {
+    if (!result.info && result.isLoading) {
+      // NOTE: .get() will trigger any actual fetches, whereas .getState() will not
+      void roomsInfoStore.get(roomId);
+    }
+  }, [roomsInfoStore, roomId, result]);
+
+  return result;
 }
 
 function useRoomInfoSuspense_withClient(client: OpaqueClient, roomId: string) {
@@ -952,7 +963,7 @@ function useUserThreads_experimental<M extends BaseMetadata>(
     },
   }
 ): ThreadsAsyncResult<M> {
-  const client = useClient<M>();
+  const client = useClient();
 
   const { store, subscribeToUserThreadsDeltaUpdates: subscribeToDeltaUpdates } =
     getLiveblocksExtrasForClient<M>(client);
@@ -1012,7 +1023,7 @@ function useUserThreadsSuspense_experimental<M extends BaseMetadata>(
     },
   }
 ): ThreadsAsyncSuccess<M> {
-  const client = useClient<M>();
+  const client = useClient();
 
   const { store } = getLiveblocksExtrasForClient<M>(client);